[FIX] web_editor, website: fix replacing a social media icon
Steps to reproduce the issue:

- Enter website edit mode.
- Drag and drop a "Social Media" snippet into the footer.
- Click on it.
- In the options, click the "Add New Social Network" button.
- Save the page.
- Re-enter edit mode.
- Click the pencil icon of the newly added item.
- In the options, click the "Replace" button.
- Bug: the media dialog does not open.

The bug was introduced by commit [1], where the double-click on the
icon, which was triggered when clicking the "Replace" button, was
replaced with a direct call to the `openMediaDialog` function.

After this change, in the steps described above, the function is called
when there is no selection on the page. As a result, `openMediaDialog`
does not execute completely.

This commit fixes the issue by selecting the icon if no selection is
already present.

[1]: https://github.com/odoo/odoo/commit/3c89439a16c41d553322893761a35592b73a5338

opw-4734855

closes odoo/odoo#211246

X-original-commit: b5a136825d8243cabefd7a96dd11214cb2cf5183
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>