[IMP] base: do not aggregate interval,priority in ir.cron
Before this commit, the interval and priority field in ir.cron were
aggregated when a group by operation was applied by getting
summed up. However, we do not get any useful information from
the sum of priorities so this so this commit removes it. It also partly
addresses an issue where the Scheduled Actions list view would
clip the grouped field's label when displaying the aggregated sum
of priorities as the first column.

opw-4536133

closes odoo/odoo#211001

X-original-commit: 6be428079fc090491b2fbf281af4f78f45a4802e
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>