[FIX] web: Odd rows on stripped lists are not highlighted
When selecting an even row by clicking on it (if it's not opening a
record) or by selecting the checkbox and then unchecking it, it will
highlight this row. This is only working on even rows but not on odd
rows.

This fix aims to make even and odd rows work correctly by being
highlighted.

task-4809597

closes odoo/odoo#210987

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>