[FIX] web: Hoot - log errors once
Before this commit, when an error occured in a test, it was logged once
(with 'console.error') when it was thrown, and then another error log
with the failed test results would be logged as well.

This meant that on the CI logs (where mostly 'console.error' logs are
displayed), there were 2 errors per test, the first of which was
irrelevant since it appeared in the second.

This commit replaces the 'console.error' of the "immediate" error logs
by a 'console.trace' (with added style to recognize the issue), to still
be able to get the traceback when developping, and to not bloat up the
logs on the CI.

Part-of: odoo/odoo#211426
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>